requirements.txt
streamlit==1.32.0
streamlit-option-menu==0.3.12
streamlit-authenticator==0.3.2
streamlit-chat==0.1.1
langchain-community==0.2.1
streamlit-pills==0.3.0
langchain==0.2.1
python-dotenv 
openai==1.30.4
pypdf==4.2.0
tiktoken==0.7.0
psycopg2-binary==2.9.9
pgvector==0.2.5
langchain-openai==0.1.7
python-magic==0.4.27
openpyxl==3.1.5
faiss-cpu